fix: layout & not found config
